% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_plot.R
\name{add_streams}
\alias{add_streams}
\title{Add pretty stream layer to ggplot2}
\usage{
add_streams(
  plot,
  stream,
  min = 0.2,
  max = 0.7,
  color = "dodgerblue4",
  min_vis = 10
)
}
\arguments{
\item{plot}{the ggplot2 object you want to add the streams to}

\item{stream}{the sf layer with streams (should be an NHDplus layer)}

\item{min}{the line size of the smallest stream}

\item{max}{the line size of the largest streams}

\item{color}{the color of the streams}

\item{min_vis}{the minimum visibility value to display (visibility column divided by 10000)}
}
\value{
a plot with the streams added
}
\description{
Uses an NHDplus stream layer to add steams to a geospatial ggplot, uses
the visibility column in the stream layer to make bigger streams thicker on
the plot.
}
