% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn-functions.R
\name{best_var}
\alias{best_var}
\title{Test multiple autocorrelation variance structures}
\usage{
best_var(
  formula,
  dataset,
  corModels = c("Exponential.tailup", "Exponential.taildown", "Exponential.Euclid"),
  addfunccol = "afvArea",
  random = NULL
)
}
\arguments{
\item{formula}{the best fitting varaibles from glmssn model}

\item{dataset}{the ssn object being modeled}

\item{corModels}{all the variance structures you want tested (tail up, tail down, euclidian)}

\item{addfunccol}{the additive function column}

\item{random}{a character vector of random effects (the column name)}
}
\value{
a printout of the fits of the different models
}
\description{
Test multiple autocorrelation variance structures
}
\examples{
\dontrun{
best_var(DOC_mgL ~ PRECIP + BFI + AWC + AREA_km2 + sev, fall_ssn)}

}
