% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn2-functions.R
\name{best_var_2}
\alias{best_var_2}
\title{Test multiple autocorrelation variance structures (fixed with SSN2)}
\usage{
best_var_2(
  formula,
  dataset,
  tailup_type = "exponential",
  taildown_type = "exponential",
  euclid_type = "exponential",
  addfunccol = "afvArea",
  random = NULL
)
}
\arguments{
\item{formula}{the best fitting varaibles from glmssn model}

\item{dataset}{the ssn object being modeled}

\item{addfunccol}{the additive function column, **currently broken, only uses afvArea}

\item{random}{a formula (~Random), with the random variables you want to add}

\item{corModels}{all the variance structures you want tested (tail up, tail down, euclidian)}
}
\value{
a printout of the fits of the different models
}
\description{
Test multiple autocorrelation variance structures (fixed with SSN2)
}
\examples{
\dontrun{
best_var(DOC_mgL ~ PRECIP + BFI + AWC + AREA_km2 + sev, fall_ssn)}

}
