% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn-functions.R
\name{best_var_str}
\alias{best_var_str}
\title{Test all potential types of variance structures}
\usage{
best_var_str(
  formula,
  dataset,
  corstr = c("tailup", "taildown", "Euclid"),
  cortype = c("Exponential", "LinearSill", "Spherical", "Mariah", "Epanech", "Cauchy",
    "Gaussian"),
  addfunccol = "afvArea"
)
}
\arguments{
\item{formula}{the best fitting varaibles from glmssn model}

\item{dataset}{the ssn object being modeled}

\item{corstr}{the types of autocorrelation to include}

\item{cortype}{the types of autocorrelation models}

\item{addfunccol}{additive function column}
}
\value{
a df with the model fits
}
\description{
Test all potential types of variance structures
}
\examples{
\dontrun{
test <- best_var_str(DOC_mgL ~ PRECIP + BFI + AWC + AREA_km2 + sev, fall_ssn,
corstr=c("tailup","taildown"))}

}
