% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn2-functions.R
\name{best_var_str_ssn2}
\alias{best_var_str_ssn2}
\title{Test all potential types of variance structures}
\usage{
best_var_str_ssn2(
  formula,
  dataset,
  tailup_type = c("none", "linear", "spherical", "exponential", "mariah", "epa"),
  taildown_type = c("none", "linear", "spherical", "exponential", "mariah", "epa"),
  euclid_type = c("none", "spherical", "exponential", "gaussian", "cosine", "cubic",
    "pentaspherical", "wave", "jbessel", "gravity", "rquad", "magnetic"),
  addfunccol = "afvArea",
  random = NULL,
  parallel = T,
  ncores = NULL
)
}
\arguments{
\item{formula}{the best fitting variables from glmssn model}

\item{dataset}{the ssn object being modeled}

\item{addfunccol}{additive function column}

\item{random}{a formula (~Random), with the random variables you want to add}

\item{corstr}{the types of autocorrelation to include}

\item{cortype}{the types of autocorrelation models}
}
\value{
a df with the model fits
}
\description{
Test all potential types of variance structures
}
\examples{
\dontrun{
test <- best_var_str(DOC_mgL ~ PRECIP + BFI + AWC + AREA_km2 + sev, fall_ssn,
corstr=c("tailup","taildown"))}

}
