% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn-functions.R
\name{burn_impact}
\alias{burn_impact}
\title{Calculate impact of burn severity}
\usage{
burn_impact(model)
}
\arguments{
\item{model}{the fitted glmssn object}
}
\value{
a table showing the range and 95% CI of burn severity for each severity level
}
\description{
uses the dNBR thresholds from the Holiday Farm fire to extract the impact of
burn severity on DOC (mg/L)
}
