% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn-functions.R
\name{check_model}
\alias{check_model}
\title{Run checks on SSN models}
\usage{
check_model(model, response)
}
\arguments{
\item{model}{the fitted glmssn model}

\item{response}{the variable of interest}
}
\value{
a list of checks (1) is the spatial residuals with just the fixed effects
(2) is the a histogram of the standardized residuals
(3) is a table of potential outlier values based on residuals and cook's distance
(4) is the cross validation plot, potential outliers are in red and numbered with rid
(5) is plots of the explanatory variables vs the standarized residuals (currently only works with continous variables)
(6) is the R-squared value for the cross validation plot
(7) is the unexplained variance (the nugget) after accounting for the spatial autocorrelation and covariates
}
\description{
Run checks on SSN models
}
