% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{clean_mtbs}
\alias{clean_mtbs}
\title{Clean MTBS wildfire files}
\usage{
clean_mtbs(wd, save_loc)
}
\arguments{
\item{wd}{the location with the raw MTBS files}

\item{save_loc}{the location to save all the files}
}
\description{
MTBS downloads come in multiple zipped folders and the file names are long and meaningless. This
will pull all files in a single folder and rename files to be more user friendly.
}
\examples{
\dontrun{
wd <- "Z:/3_GIS-GPS/Wildfire/MTBS/McKenzie/"
save_loc <- "Z:/3_GIS-GPS/Wildfire/MTBS/McKenzie"
clean_mtbs(wd, save_loc)}
}
