% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_photos.R
\name{clean_photo_names}
\alias{clean_photo_names}
\title{Clean field photo names}
\usage{
clean_photo_names(wd, datasheet, group = NA)
}
\arguments{
\item{wd}{the file path where the photos are located written as a string}

\item{datasheet}{a dataframe of the datasheet}

\item{group}{string with the name of the group (see details)}
}
\value{
The photos in the same place just renamed with the site and time.
}
\description{
The group name is used to subset the datasheet if there are
multiple groups on the datasheet. This should be a string.
Datasheet should have the following columns: Group (if grouping is used),
Sample_Time_PST, Site_Num
}
\details{
Uses the times sites were visited from the datasheet and the metadata from
the photos to rename confusing/undescriptive names from phones and tablets
to the site it was taken at with the data and time the photo was taken.
}
\examples{
\dontrun{
date <- "2022_11_01"
group <- "Group 1"
wd <- paste("Z:/1_Research/2_PNNL_EWEB/Photos", date, group, sep="/")
datasheet <- read.csv(paste("Z:/1_Research/2_PNNL_EWEB/Field Notes/",
date, "/Master_Sample_List_", date, ".csv", sep=""))
clean_photo_names(wd, datasheet, group="Group1")}
}
