% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{clean_raster}
\alias{clean_raster}
\title{Clip raster to shapefile}
\usage{
clean_raster(raster, sf, type = "numeric", res = NULL, return = "df")
}
\arguments{
\item{raster}{the raster you want to clip}

\item{sf}{the shapefile you want to use to clip the raster}

\item{type}{either 'numeric' or 'categorical' depending on if your raster is discrete or continuous values}

\item{res}{the resolution of the projected raster, if not specified with default to 30m}

\item{return}{either 'df' or 'raster' to specify the form of the returned raster}
}
\value{
if 'return' is df it will return the raster as a dataframe suitable for
plotting in ggplot2. If 'return' is raster it will return the raster as a
rasterLayer object
}
\description{
Function will ensure the raster and shapefile are in the same coordinate reference
system, then will clip to the shapefile boundry.
}
