% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{convert_crs}
\alias{convert_crs}
\title{Convert the projection geospatial data}
\usage{
convert_crs(input, goal, type = "numeric", res = NULL)
}
\arguments{
\item{input}{the shapefile or raster you want to convert to a new coordinate reference system}

\item{goal}{the shapefile or raster with the coordinate reference system desired}

\item{type}{either 'numeric' or 'categorical' depending on if your raster is discrete or continuous values}

\item{res}{the resolution of the projected raster, if not specified with default to 30m}
}
\value{
the input object with the coordinate reference system of the goal object
}
\description{
Converts a raster or shapefile into the same coordinate reference system as
another raster or shapefile.Helpful when trying to use multiple datasets.
}
\examples{
\dontrun{
convert_crs(basin, DEM)
}
}
