% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_date}
\alias{file_date}
\title{Format date for using in file names}
\usage{
file_date(date = Sys.Date(), yeardigit = 4)
}
\arguments{
\item{date}{a date in the form YYYY-MM-DD, default is today's date}

\item{yeardigit}{either 2 or 4 to indicate number of year digits desired}
}
\value{
a string of the date with underscores between the year, month, and day
}
\description{
Format date for using in file names
}
\examples{
file_date("2023-02-13")
}
