% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{per_change}
\alias{per_change}
\title{Calculate percent change}
\usage{
per_change(old, new)
}
\arguments{
\item{old}{the original value}

\item{new}{the new value}
}
\value{
the percent change between the old and new values as a percent (%)
}
\description{
Calculate percent change
}
\examples{
per_change(1,2)

}
