% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn2-functions.R
\name{plot_ssn2}
\alias{plot_ssn2}
\title{plot SSN data (updated for SSN2)}
\usage{
plot_ssn2(
  ssn_obj,
  sites = c("sites", "preds"),
  type = c("obs", "pred"),
  response = c("DOC_mgL", "DOC_mgL"),
  size = c(2, "se"),
  scale = "binned",
  nbins = 8,
  palette = "cubicl",
  prec = 2,
  bin_method = "jenks"
)
}
\arguments{
\item{ssn_obj}{a list or individual ssn_glm or ssn_lm object}

\item{sites}{a vector of sites names, should match the ssn.ssn file names}

\item{type}{a vector of the type of data being plotted, should be either "obs" or "pred"}

\item{response}{the variable to plot as a character}

\item{size}{the size to plot to the variable, a vector the same length as models, either a number or "se}

\item{scale}{Either "continuous" or "binned"}

\item{nbins}{the number of bins to use in plotting}

\item{palette}{a character with the color palette to use. Currently 'parula', 'ocean.haline', 'cubicl', and 'kovesi.rainbow' from the 'pals' package are supported (and custom 'katie_pal')}

\item{prec}{an integer for the number of significant figures used for binning the intensity}

\item{bin_method}{either "jenks" or "quantile"}
}
\value{
a list of the items needed to make a nice clean map of the data
}
\description{
takes either a glmssn fit or a glmssn.predict as an input and returns a list of the
items needed to make a nice clean map of the data. If you use multiple datasets it
will set them on a single scale for comparison accross the maps.
}
