% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn-functions.R
\name{plot_ssn_resid}
\alias{plot_ssn_resid}
\title{Plot glmssn residuals in ggplot2}
\usage{
plot_ssn_resid(model, sites = "sites.shp", size = "_resid_")
}
\arguments{
\item{model}{the glmssn object}

\item{sites}{the shapefile name for the sites in the .ssn object}

\item{size}{the column used to determine the size of the points}
}
\value{
a ggplot object of residuals on the stream network
}
\description{
Given a glmssn object, it will extract the residuals and plot as an
ggplot
}
