% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{samp_date}
\alias{samp_date}
\title{Convert date to one used for samples}
\usage{
samp_date(date = Sys.Date(), yeardigit = 2)
}
\arguments{
\item{date}{a date in the form YYYY-MM-DD, default is today's date}

\item{yeardigit}{either 2 or 4 to indicate number of year digits desired}
}
\value{
a character in the form of YYMMDD or YYYYMMDD used for logging samples
}
\description{
Convert date to one used for samples
}
\examples{
samp_date("2023-02-13")
}
