% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{season}
\alias{season}
\title{Get season from date}
\usage{
season(Date)
}
\arguments{
\item{Date}{a vector of dates}
}
\value{
a vector of seasons
}
\description{
Spring is defined as months March through May, Summer is months June through
August, Fall is September through November, and Winter is Decemeber through
February.
}
\examples{
dates <- c("2023-02-13", "2023-07-16")
as.Date(dates)
season(dates)
}
