% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{summary_raster}
\alias{summary_raster}
\title{Summarise raster dataset after clipping to a basin shape}
\usage{
summary_raster(
  raster,
  basin = NULL,
  type = "numeric",
  return = "t",
  data_type = "character"
)
}
\arguments{
\item{raster}{object of class raster}

\item{basin}{object of class sf, a shapefile used to clip the raster optional}

\item{type}{either 'categorical' or 'numeric'}

\item{return}{either 'p' for plot or 't' for table (see details for more info)}

\item{data_type}{either 'character' or 'wildfire' (see details for more info)}
}
\value{
if 'type' is numeric will provide mean, median, min, max, if categorical will return table with percent as a decimal for each category
}
\description{
Summarise raster dataset after clipping to a basin shape
}
\details{
If data_type is 'character' it will return a table with either percentage of total in the basin
or if it's a numeric it will give quantiles to describe the range of values.
If the data_type is 'wildfire' it will return the percent burned at each severity
with 1 being unburned to low severity and 4 being high severity. If return is 'p', it will
provide a bar plot showing the relative counts of each group, 't' returns a table with
that data.
}
