% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_green}
\alias{theme_green}
\title{Katie's personal theme}
\usage{
theme_green()
}
\description{
Keeps with the green theme branding of my work. Uses Barlow Semi Condensed font
and used the green colors from google slides.
}
\details{
Make sure the font is installed from here: (https://fonts.google.com/specimen/Barlow+Semi+Condensed)
Then Just once you'll need to run the following code:
font_add("Barlow Semi Condensed", "BarlowSemiCondensed-Regular.ttf")
font_add("Barlow Semi Condensed ExtraBold", "BarlowSemiCondensed-ExtraBold.ttf")
font_add("Barlow Semi Condensed SemiBold", "BarlowSemiCondensed-SemiBold.ttf")
}
