% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_pub}
\alias{theme_pub}
\title{ggplot theme for publication figures}
\usage{
theme_pub()
}
\description{
Used to help ensure consistent formatting across publication figures. Based on
ggclean with a few modifications.
}
\examples{
x <- sample(1:250, 50)
y <- sample(c("A","B", "C"), 50, replace=T)
z <- sample(c("LOW","MED","HIGH"), 50, replace=T)
w <- sample(1:250, 50)
df <- data.frame(x=x, y=y, z=z, w=w)

ggplot(df, aes(x=y, y=x, fill=z)) + geom_boxplot() + theme_pub()
ggplot(df, aes(x=x, y=w, color=y)) + geom_point() + geom_smooth(method="lm") + theme_pub()
}
