% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn2-functions.R
\name{var_imp2}
\alias{var_imp2}
\title{Plot and report variable importance based on standarized linear regression coefficients}
\usage{
var_imp2(model)
}
\arguments{
\item{model}{the fitted glmssn object}
}
\value{
A list, with the first object being a table of the coefficients and the second
being a plot of the coefficients. Includes the standard error bars. If you did model
selection these are likely larger than is shown, but you can still show, just make it
very clear in the paper that these are only error from the final model.
}
\description{
Plot and report variable importance based on standarized linear regression coefficients
}
