% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_pub}
\alias{theme_pub}
\title{ggplot theme for publication figures}
\usage{
theme_pub()
}
\description{
Used to help ensure consistent formatting across publication figures. Based on
ggclean with a few modifications.
}
